% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{A549ChromLoops}
\alias{A549ChromLoops}
\title{Predicted chromsome loops in A549's chromsome 18}
\format{A dataframewith 216 rows and 3 columns:
\describe{
  \item{chr}{chromosome 18}
  \item{start}{chromatin loops' start coordinates}
  \item{stop}{chromatin loops' end coordinates}
}}
\source{
\url{http://promoter.bx.psu.edu/hi-c/publications.html}
}
\usage{
A549ChromLoops
}
\description{
A subset 216-by-3 dataframe containing predicted and Peackahu-merged chromatin loops in chromsome 18 of A549 cells.
}
\references{
{
Salameh TJ, Wang X, Song F, Zhang B, Wright SM, Khunsriraksakul C, Yue F. (2019) A supervised learning framework for chromatin loop detection in genome-wide contact maps. \emph{biorXiv}. doi: \url{https://doi.org/10.1101/739698}
}
}
\keyword{datasets}
