% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{antrack}
\alias{antrack}
\title{A Classified Annotation Track of H3K9me3 predictor in mm10 chromosome 18 loops}
\format{An object of class \code{GeneRegionTrack} of length 25.}
\usage{
antrack
}
\description{
A sample clustered annotation track generated by the classifiedAnnotationTrack function.
}
\keyword{datasets}
