% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{bin_motifs}
\alias{bin_motifs}
\title{Calculate Motif Membership Probabilty per TAD}
\usage{
bin_motifs(motifs, tad, j, M, offset)
}
\arguments{
\item{motifs}{A dataframe of matched motifs with at least the following attributes:
\itemize{
\item \emph{start}: an integer denoting the starting coordinate of matched motif
\item \emph{stop}: an integer denoting the end coordinate of matched motif
}}

\item{tad}{A row of a dataframe containing hierarchically clustered TADs. Compatible with output of the annotator3D::cluster function. The row should have at least the following attributes:
\itemize{
\item \emph{segment1}: an integer denoting the starting coordinate of the predicted TAD
\item \emph{segment2}: an integer denoting the end coordinate of the predicted TAD
}}

\item{j}{An integer denoting the index of the TAD for classification}

\item{M}{A n-by-m numeric matrix to be used as a dynamic array of one TAD.}

\item{offset}{A numeric denoting the resolution of Hi-C data bins.}
}
\value{
M A n-by-m numeric matrix storing calculated membership scores.
}
\description{
Assuming equal chance of assignment given the matched sequence, this function calculates membership
proportion of motifs per TAD.
}
