% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_visualize.R
\name{classifiedAnnotationTrack}
\alias{classifiedAnnotationTrack}
\title{Title}
\usage{
classifiedAnnotationTrack(motif_bed, gen, thresh = 0)
}
\arguments{
\item{motif_bed}{A string of the full path to the BED file
containing the classified motifs}

\item{gen}{A named character containing the chromosome name}

\item{thresh}{An optional numeric value denoting the threshold of
significance. Motifs with membership scores greater than or equal to thresh
will be returned.}
}
\value{
A Formal Class AnnotationTrack for enriched motif with GRanges
corresponding to motif's genomic ranges,
}
\description{
Title
}
