% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalCluster.R
\name{cluster}
\alias{cluster}
\title{Title}
\usage{
cluster(data, cluster_method = "ward.D2")
}
\arguments{
\item{data}{}

\item{cluster_method}{}
}
\value{
dataframe
}
\description{
Title
}
\examples{
data.cl <- cluster(data)

}
