% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_visualize.R
\name{filter_motifs}
\alias{filter_motifs}
\title{Filter BED Dataframe by Threshold}
\usage{
filter_motifs(motifs, thresh = 0)
}
\arguments{
\item{motifs}{A dataframe with column names corresponding to BED file format.}

\item{thresh}{An optional numeric value that serves as lower bound for scores
in the dataframe}
}
\value{
A dataframe with column names corresponding to BED format and scores
greater than or equal to lower bound threshold.
}
\description{
Filter classified motifs based on a lower bound threshold on their membership
score.
}
