% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/hsa.R
\docType{data}
\name{hsa}
\alias{hsa}
\title{Chromsome interactions for 40kb segments of mES cells' chromsome 18}
\format{A matrix with 1000 rows and 1000 columns:
\describe{
  \item{row}{chromosome 18 fragments start coordinates}
  \item{column}{chromosome 18 fragments end coordinates}
}}
\source{
\url{http://sysbio.rnet.missouri.edu/3dgenome/GSDB/details.php?id=TE1402WS}
}
\usage{
hsa
}
\description{
A subset 1000-by-1000 symmetric matrix of interchromsome interactions murine chromsome 18 embryonic stem cells. Murine genome was sequenced on Illumina HiSeq 2000 after HindIII and NCoI digestion, achiving resolution of 40 kb.
}
\references{
{
Dixon JR, Selvaraj S, Yue F, Kim A \emph{et al.} (2012) Topological domains in mammalian genomes identified by analysis of chromatin interactions. \emph{Nature} \strong{485}(7398):376--80. PMID: \href{https://www.ncbi.nlm.nih.gov/pubmed/22495300}{22495300}
}
}
\keyword{datasets}
