% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/datasetParse.R
\name{LinearizeNestedList}
\alias{LinearizeNestedList}
\title{Title}
\usage{
LinearizeNestedList(NList, LinearizeDataFrames = FALSE, NameSep = "/",
  ForceNames = FALSE)
}
\arguments{
\item{NList}{}

\item{LinearizeDataFrames}{}

\item{NameSep}{}

\item{ForceNames}{}
}
\value{
list
}
\description{
Title
}
