% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralSegmentation.R
\name{mutual_knn_graph}
\alias{mutual_knn_graph}
\title{Nearest neighbourdetection based on Euclidean distances}
\usage{
mutual_knn_graph(X, nn = 10)
}
\arguments{
\item{nn}{number of k-nearest neighbours to consider}

\item{x}{the chromatin interaction/contact matrix}
}
\value{
matrix
}
\description{
Finds nn nearest Euclidean neighbours. This function is authored by Nura Kawa (2018).
}
\references{
{\cite{Kawa, N. 2018. Spectral Clustering. RPubs. URL: \url{https://rpubs.com/nurakawa/spectral-clustering}}}
}
\author{
Nura Kawa
}
