% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_visualize.R
\name{plot_layered_gviz}
\alias{plot_layered_gviz}
\title{Plot a Layered Gviz Figure}
\usage{
plot_layered_gviz(antrack, gen, from = 1, to = 1e+06)
}
\arguments{
\item{antrack}{A classified annotation track obtained with
\code{classifiedAnnotationTrack}.}

\item{gen}{A named character containing genome of interest}

\item{from}{An optional integer value of start of viewing window.}

\item{to}{An optional integer value of end of viewing window.}
}
\description{
Plot the classified motifs, along with the ideogram of the chromosome of
interest and and genome axis.
}
\examples{
\dontrun{
data("A549ChromLoops")
data("matchedMotifs")
classify(matchedMotifs, A549ChromLoops)
gen <- c("mm10")
names(gen) <- "chr18"
antrack <- classifiedAnnotationTrack("./BEDs/H3K9me3.bed", gen)
plot_layered_gviz(antrack, gen)
}

}
