% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runAnnotationApp.R
\name{runAnnotator}
\alias{runAnnotator}
\title{Run annotator3D as a Shiny app}
\usage{
runAnnotator()
}
\description{
runAnnotationApp runs annotator3D as a Shiny app.
The user needs to upload FIMO matched motif TSV file and a TSV file
containing predicted chromatin loops. The user can then customize parameters
of visualization such as lower bound for loop membership score, and Gviz
plotting options.
}
\details{
Example TSV files are included in \code{extdata} to run annotator3D.
}
\examples{
\dontrun{
  runAnnotator()
}

}
