% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralSegmentation.R
\name{spectral_clustering}
\alias{spectral_clustering}
\title{Spectral segmentation of chromatin interaction/contact matrices}
\usage{
spectral_clustering(x, nn = 10, n_eig = 2)
}
\arguments{
\item{x}{the chromatin interaction/contact matrix}

\item{nn}{number of k-nearest neighbours to consider}

\item{n_eig}{number of eigenvectors to keep}
}
\value{
eigenvectors of clustered components in X for
 kmeans analysis
}
\description{
Creates eigenvectors of clustered components in the chromatin interaction/contact matrix. This function is authored by Nura Kawa (2018).
}
\examples{
clustered <- spectral_clustering(X)

}
\references{
{\cite{Kawa, N. 2018. Spectral Clustering. RPubs. URL: \url{https://rpubs.com/nurakawa/spectral-clustering}}}
}
\author{
Nura Kawa
}
