% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/optimalCluster.R
\name{tad.group}
\alias{tad.group}
\title{Title}
\usage{
tad.group(dendrogram, tads, k, plot_subtads = FALSE)
}
\arguments{
\item{dendrogram}{}

\item{tads}{}

\item{k}{}

\item{plot_subtads}{}
}
\value{
dataframe
}
\description{
Title
}
