% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/classification.R
\name{write_to_BED}
\alias{write_to_BED}
\title{Write Motif Vector to BED File}
\usage{
write_to_BED(motif_ranges, motif_name, chrom)
}
\arguments{
\item{motif_ranges}{A numeric vector of motif's membership scores. The names
of the entries correspond to the motif's location on the chromsome chrom.}

\item{motif_name}{A string containing the name of the motif to be stored.}

\item{chrom}{A string denoting the chromosome containing the motif and
chromatin loop.}
}
\description{
Write a motif and its location to a BED file stored locally.
}
