% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/annotation_visualize.R
\name{classifiedAnnotationTrack}
\alias{classifiedAnnotationTrack}
\title{Get Classified Motif Annotation Track}
\usage{
classifiedAnnotationTrack(motif_bed, gen, thresh = 0)
}
\arguments{
\item{motif_bed}{A string of the full path to the BED file
containing the classified motifs}

\item{gen}{A named character containing the chromosome name}

\item{thresh}{An optional numeric value denoting the threshold of
significance. Motifs with membership scores greater than or equal to thresh
will be returned.}
}
\value{
A Formal Class GeneRegionTrack for enriched motif with GRanges
corresponding to motif's genomic ranges,
}
\description{
This function returns an annotated GeneRegionTrack that can be plotted alone
or layered with other Gvix/ggbio tracks.
}
\examples{
\dontrun{
data("A549ChromLoops")
data("matchedMotifs")
classify(matchedMotifs, A549ChromLoops)
gen <- c("mm10")
names(gen) <- "chr18"
antrack <- classifiedAnnotationTrack("./BEDs/H3K9me3.bed", gen)
}

}
