% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/spectralSegmentation.R
\name{graph_laplacian}
\alias{graph_laplacian}
\title{Laplacian of contact matrix}
\usage{
graph_laplacian(W)
}
\arguments{
\item{W}{similarity matrix}
}
\value{
matrix Laplacian transformation of the contact matrix
}
\description{
Calculates the Laplacian transformation of the contact matrix. This function is authored by Nura Kawa (2018) and modified by Nada Elnour (2019).
}
\references{
{\cite{Kawa, N. 2018. Spectral Clustering. RPubs. URL: \url{https://rpubs.com/nurakawa/spectral-clustering}}}
}
\author{
Nura Kawa

Nada Elnour
}
