% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/LinearizeNestedList.R
\name{LinearizeNestedList}
\alias{LinearizeNestedList}
\title{Flatten Nested Lists to Get Index  Paths to Items}
\source{
\url{https://github.com/mrdwab/mrdwabmisc/tree/master/R}
}
\usage{
LinearizeNestedList(NList, LinearizeDataFrames = FALSE, NameSep = "/",
  ForceNames = FALSE)
}
\arguments{
\item{NList}{A list of objects.}

\item{LinearizeDataFrames}{A boolean denoting whether NList is a data frame to be converted to a list.}

\item{NameSep}{A character that serves as separator between levels of nested list}

\item{ForceNames}{A boolean denoting whether LinearizeNestedList should return a named path to items in NList.}
}
\value{
LinearList: Flat named list
}
\description{
Recursively linearize a nested list. If the list is flat, do nothing.
}
\references{
{
Mahto A Turner R, Bhel AS. (2016) mrdwabmisc: Miscellaneous R functions, mostly for data processing.
}
}
\author{
Akhil S. Bhel
}
