% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/clustering.R
\name{tad.group}
\alias{tad.group}
\title{Label TADs into Flat and Hierarchical Clusters}
\usage{
tad.group(dendrogram, tads, k, plot_subtads, offset)
}
\arguments{
\item{dendrogram}{A dendrogram (list) produced by hclust.}

\item{tads}{A cut dendrogram object into $$k$$ objects}

\item{k}{The number of groups into which the dendrogram is cut}

\item{plot_subtads}{A boolean specifying whether sub-TADs should be plotted}

\item{offset}{An integer denoting the binning resolution of the Hi-C dataset}
}
\value{
A dataframe with four attributes:
\itemize{
\item \emph{segment1}: the start coordinate of a Hi-C bin
\item \emph{segment2}: the end coordinate of a Hi-C bin
\item \emph{flat_cluster}: predicted cluster group of TADs output by hcluster
\item \emph{hcluster}: full hierarchical subclusters of the TAD given the flat_cluster
}
}
\description{
Finds and groups sub-TADs and TADs into hierarchical clusters, with the option of the plotting them separatelyy.
}
